Add PrettyPrinter
The stateful design (compared to free functions) allows
future optimizations, e.g: storing and re-using the
result of nsSinceEpochToBrokenDownTimeUTC,
as the time difference between events is usually small.

The logic of substituting %x markers in format strings
is duplicated in printEvent and printTime. Duplication
could be avoided by making a template,
taking a set of arguments and calling an overloaded
member with them, but overall, the code becomes harder
to follow.