BINLOG_CREATE_SOURCE_AND_EVENT: swap clock and format args
The last named argument of the macro is mixed with the
args... argument, to prevent __VA_ARGS__ becoming empty
when the log event has no arguments.

Mixing clock with args... is a logical choice, as it yields
to less code, since SessionWriter::addEvent takes clock
before args..., and clock is otherwise not used in the macro.

However, higher level macros (to be introduced later)
would have only format and args... parameters, leaving
no room for choosing which parameter to mix with __VA_ARGS__,
and I could not find a way to change the first element of
the variadic pack, or insert a new element after the first one.