Rename Actor to WriterProp
Actor is never mentioned at the Session/SessionWriter level,
which allows to set the id/name of SessionWriter.
It is surprising that the name set to the writer
appears as the name of the 'actor' while pretty printing.
To keep the glossary of the project minimal,
the concept `Actor` is dropped.

Actor also means various things in various contexts elsewhere,
which might confuse people.