Specify integration test build dir in terms of target location
After mkdir Release && cd Release && cmake .. :
PROJECT_BINARY_DIR expands to src/Release, but on
windows, when generating visual studio solution files,
the binaries end up in src/Release/Release, as
visual studio supports multiple configuration...

CMAKE_CFG_INTDIR supposed to help here,
as on single configuration platforms (make)
it expands to `.`, but with visual studio,
it expands to $(Configuration) or $(OutDir),
depending on the VS version.
These expressions are replaced by VS,
but we want to use them in the ctest
context, where they are used verbatim.

The current solution is to take the directory
of the IntegrationTest target, which will
be the same for each integration test program.