Add GitHub workflow to build each PR on 3 platforms
Compile on linux+GCC, macOS+Clang, windows+MSVC.
Test on linux and windows.

Do not test on macOS, as it does not have boost,
and installing it requires a complete boost build
from source, that takes too much time.

On linux, boost libraries have to be installed
from the distro repository, as the container does
not have them installed:
https://github.com/actions/virtual-environments/issues/83

The installed versioned libs must be symlinked
to non-versioned names, otherwise cmake will not find them.

On windows, BOOST_ROOT must be added to the path,
otherwise the tests will not find the dynamic libraries.
The error message in that case:
Exit code 0xc0000135 ***Exception: