Add EntryStream - an abstraction to read entries
Make EventStream.nextEvent take an EntryStream
instead of a std::istream. This allows different
kinds of inputs (e.g: memory buffers),
not just files - useful to convert entries
to text right after consuming them from the Session.
In theory, the istream interface could allow
the same, but it was just too complicated to be useful.

The logic of constructing an entry (size+payload)
from a stream of bytes was moved from
EventStream to EntryStream.

The bread performance is not affected as measured,
despite the extra indirection of the virtual
nextEntryPayload call.
Performance can be further improved later,
e.g: by replacing istream by FILE*.

Range: add operator bool to allow:

    while (Range payload = input.nextEntryPayload())