Make Channel co-owned by Session and SessionWriter
Session stores shared_ptr of Channels,
shared with the respective SessionWriters.

The Channel is closed if only the Session owns it,
checked by shared_ptr::use_count.

Stability of Channel references is provided by the shared_ptr,
storing them in a list is no longer necessary -
it is replaced with a vector, that is faster to iterate.

Motivation: use-after-free if SessionWriter is destroyed
later than the Session it references. See contrived
test swap_writers_of_different_sessions that reproduces the issue.