Disable clang-tidy on IntegrationTest.cpp
clang-tidy is triggered by a Boost.Process error:
https://github.com/boostorg/process/issues/184

Regardless that boost is included with -isystem,
and also excluded via the HeaderFilterRegex,
it produces this error:

boost/range/as_literal.hpp:43:20: error: Null pointer passed as an argument to a 'nonnull' parameter

Perhaps it thinks that because of the BOOST_AUTO_TEST_CASE macro,
the code is in our source file.
If // NOLINT is added to the calling line, the error becomes
incomprehensible, but it remains there.