Allow logging pointer addresses, display them in hex
The C++20 specified remove_cvref does not remove the `const` from
a `const void*`, therefore we need one CustomSerializer and
CustomTag for both `void*` and `const void*`.

Logging addresses as uintptr_t was considered,
the advantage being saved space on x32.
However, that requires the ToStringVisitor to do a bit more work
(ensure that the field tag is valid, fetch a tag specific
integral depending on the field type)
- and the extra complexity was too high.