Store EventSources in the binary compile-time
Instead of creating an EventSource object the first
time a log macro is executed, persist EventSource details
compile time into custom ELF sections, and extract
them in the Session on the first consume.

This avoids an extra branch for each log event,
and a huge latency spike for the first execution of each event source.

This change temporarily breaks GCC compatibility.
This change is not compatible with macOS or windows.