Implement static storage on macOS
Pointers in Mach-O binaries are rebased by the dynamic loader
during loading, because of ASLR. Rebasing need to be done
differently on different macOS versions. To avoid this complication,
we parse the loaded images instead, that has the pointers
already rebased. This also makes determining the event source id
simpler.

TestSessionWriter: the minimum queue size was quadrupled
on M1 because of the different page size. Without the metadata
going through the queues (this change), the old event count
does not trigger a new queue allocation: increase the event count.