Fix event_sources_of_running_program on linux+lld
In our custom ELF section, .binlog.esrc, we are storing
8 pointers/EventSource, see create_source.hpp.

Previously, we listed the loaded files,
and extracted these pointers from the on-disk files
where the program was loaded from.

GNU ld and gold sets these pointers build time
to offsets in the file, that are relocated load time.
lld, however, leaves them blank, a different kind
of relocation is used instead.

Therefore, we have to get the sources from memory, instead of disk.
We are using dl_iterate_phdr to get the list of loaded
objects + their load address. The program headers
reference segments, a set of mapped sections.
Unfortunately, I wasn't able to find the
section headers in memory (probably it isn't there),
so we still need to read the original file, find our custom
section, and map it to a loaded segment.

Since we are reading the pointers from loaded memory,
our section must be "allocated". Since it contains relocations,
it must be marked as writable, otherwise the linker warns:
"relocation in read-only section `.binlog.esrc'".
Therefore, we add "wa" to the .pushsection directive.

A range-check bug in the original implementation was found
and fixed by Eric Harding, that triggered the investigation
that ended with this fix - thanks!